requirements.txt
asgiref==3.5.1
Django==3.2.10
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
